Amélioration de l'affichage

+++ view/match_view.py
@@ -11,8 +11,8 @@ class MatchView:
         print(f"{match[0].prenom} Vs {match[1].prenom}\n")
 
     def enterMatchResults(self, listPairs):
-        player_score1 = int(input(f"* Veuillez saisir le score du joueur {listPairs[0].prenom}: "))
-        player_score2 = int(input(f"* Veuillez saisir le score du joueur {listPairs[1].prenom}: "))
+        player_score1 = float(input(f"* Veuillez saisir le score du joueur {listPairs[0].prenom}: "))
+        player_score2 = float(input(f"* Veuillez saisir le score du joueur {listPairs[1].prenom}: "))
         listPairs[0].score.append(player_score1)
         listPairs[1].score.append(player_score2)
         liste = (listPairs[0].prenom, (player_score1)), (listPairs[1].prenom, (player_score2))
